+++ src/modules/common/rawstr.cpp
@@ -183,9 +183,9 @@ signed char RawStr::findOffset(const char *ikey, __u32 *start, __u16 *size, long
 			headoff = 0;
 
 			stdstr(&key, ikey, 3);
-			if (!caseSensitive) toupperstr_utf8(key, strlen(key)*3);
+			if (!caseSensitive) toupperstr_utf8(key, (unsigned int)(strlen(key)*3));
 
-			int keylen = strlen(key);
+			int keylen = (int)strlen(key);
 			bool substr = false;
 
 			trybuf = maxbuf = 0;
@@ -243,7 +243,7 @@ signed char RawStr::findOffset(const char *ikey, __u32 *start, __u16 *size, long
 		idxfd->read(&tmpStart, 4);
 		idxfd->read(&tmpSize, 2);
 		if (idxoff)
-			*idxoff = tryoff;
+			*idxoff = (__u32)tryoff;
 
 		*start = swordtoarch32(tmpStart);
 		*size  = swordtoarch16(tmpSize);
@@ -262,17 +262,17 @@ signed char RawStr::findOffset(const char *ikey, __u32 *start, __u16 *size, long
 			if (bad) {
 				if(!awayFromSubstrCheck)
 					retval = -1;
-				*start = laststart;
+				*start = (__u32)laststart;
 				*size = lastsize;
 				tryoff = lasttry;
 				if (idxoff)
-					*idxoff = tryoff;
+					*idxoff = (__u32)tryoff;
 				break;
 			}
 			idxfd->read(&tmpStart, 4);
 			idxfd->read(&tmpSize, 2);
 			if (idxoff)
-				*idxoff = tryoff;
+				*idxoff = (__u32)tryoff;
 
 			*start = swordtoarch32(tmpStart);
 			*size  = swordtoarch16(tmpSize);
@@ -346,7 +346,7 @@ void RawStr::readText(__u32 istart, __u16 *isize, char **idxbuf, SWBuf &buf) con
 	while (true);	// while we're resolving links
 
 	if (idxbuflocal) {
-		int localsize = strlen(idxbuflocal);
+		int localsize = (int)strlen(idxbuflocal);
 		localsize = (localsize < (*isize - 1)) ? localsize : (*isize - 1);
 		strncpy(*idxbuf, idxbuflocal, localsize);
 		(*idxbuf)[localsize] = 0;
@@ -381,7 +381,7 @@ void RawStr::doSetText(const char *ikey, const char *buf, long len)
 
 	char errorStatus = findOffset(ikey, &start, &size, 0, &idxoff);
 	stdstr(&key, ikey, 2);
-	if (!caseSensitive) toupperstr_utf8(key, strlen(key)*2);
+	if (!caseSensitive) toupperstr_utf8(key, (unsigned int)(strlen(key)*2));
 
 	len = (len < 0) ? strlen(buf) : len;
 
@@ -424,7 +424,7 @@ void RawStr::doSetText(const char *ikey, const char *buf, long len)
 		while (true);	// while we're resolving links
 	}
 
-	endoff = idxfd->seek(0, SEEK_END);
+	endoff = (__u32)idxfd->seek(0, SEEK_END);
 
 	shiftSize = endoff - idxoff;
 
@@ -440,7 +440,7 @@ void RawStr::doSetText(const char *ikey, const char *buf, long len)
 	memcpy(outbuf + size, buf, len);
 	size = outsize = size + (len);
 
-	start = outstart = datfd->seek(0, SEEK_END);
+	start = outstart = (__u32)datfd->seek(0, SEEK_END);
 
 	outstart = archtosword32(start);
 	outsize  = archtosword16(size);
